feat(auth): implement sign-in functionality with Mailgun provider and update environment variables
