--- src/app/auth/login/page.tsx
@@ -1,74 +1,0 @@
-"use client"
-
-import { zodResolver } from "@hookform/resolvers/zod"
-import { Suspense } from "react"
-import { useForm } from "react-hook-form"
-import { z } from "zod"
-import { Button } from "~/app/components/ui/button"
-import {
-    Form,
-    FormControl,
-    FormDescription,
-    FormField,
-    FormItem,
-    FormLabel,
-    FormMessage,
-} from "~/app/components/ui/form"
-import { Input } from "~/app/components/ui/input"
-
-const formSchema = z.object({
-    username: z.string().min(2, {
-        message: "Username must be at least 2 characters.",
-    }),
-})
-
-export function AuthPage() {
-
-    const form = useForm<z.infer<typeof formSchema>>({
-        resolver: zodResolver(formSchema),
-        defaultValues: {
-            username: "",
-        },
-    })
-
-    // 2. Define a submit handler.
-    function onSubmit(values: z.infer<typeof formSchema>) {
-        // Do something with the form values.
-        // ✅ This will be type-safe and validated.
-        console.log(values)
-    }
-
-    return (
-        <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-                <FormField
-                    control={form.control}
-                    name="username"
-                    render={({ field }) => (
-                        <FormItem>
-                            <FormLabel>Username</FormLabel>
-                            <FormControl>
-                                <Input placeholder="shadcn" {...field} />
-                            </FormControl>
-                            <FormDescription>
-                                This is your public display name.
-                            </FormDescription>
-                            <FormMessage />
-                        </FormItem>
-                    )}
-                />
-                <Button type="submit">Submit</Button>
-            </form>
-        </Form>
-    );
-}
-
-export default function Page() {
-    return (
-        <div className='flex justify-center items-center'>
-            <Suspense fallback={<div>Loading...</div>}>
-                <AuthPage />
-            </Suspense>
-        </div>
-    );
-}
