+++ src/server/auth/config.ts
@@ -1,6 +1,7 @@
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import { type DefaultSession, type NextAuthConfig } from "next-auth";
 import GoogleProvider from "next-auth/providers/google";
+import MailgunProvider from "next-auth/providers/mailgun";
 
 import { db } from "~/server/db";
 
@@ -33,6 +34,7 @@ declare module "next-auth" {
 export const authConfig = {
   providers: [
     GoogleProvider,
+    MailgunProvider,
   ],
   adapter: PrismaAdapter(db),
   pages: {
